moved to optogenetics module
